[ENH] Patchwork on NaN detect-and-retrieve scheme

+++ pynets/stats/benchmarking.py
@@ -5,6 +5,7 @@ Created on Tue Nov  7 10:40:07 2017
 Copyright (C) 2017
 @authors: Derek Pisner
 """
+import os
 from sklearn.metrics.pairwise import (
     cosine_distances,
     haversine_distances,
@@ -240,9 +241,58 @@ def _discr_rdf(dissimilarities, labels):
     return out
 
 
-def reshape_graphs(graphs):
-    n, v1, v2 = np.shape(graphs)
-    return np.reshape(graphs, (n, v1 * v2))
+def beta_lin_comb(beta, GVDAT, meta):
+    """
+    This function calculates linear combinations of graph vectors stored in
+    GVDAT for all subjects and all sessions given the weights vector beta.
+    This was adapted from a function of the same name, written by Kamil Bonna,
+    10.09.2018.
+
+    Parameters
+    ----------
+    beta : list
+        List of metaparameter weights.
+    GVDAT : ndarray
+        5d data structure storing graph vectors.
+
+    Returns
+    -------
+    gv_array : ndarray
+        2d array of aggregated graph vectors for all sessions, all subjects.
+    """
+    import numpy as np
+    import math
+
+    def normalize_beta(beta):
+        sum_weight = sum([b1 * b2 * b3 for b1 in beta[:N_atl] for b2 in
+                          beta[N_atl:N_atl + N_mod] for b3 in beta[-N_thr:]])
+        return [b / math.pow(sum_weight, 1 / 3) for b in beta]
+
+    # Dataset dimensionality
+    N_sub = meta['N_sub']
+    N_ses = meta['N_ses']
+    N_gvm = meta['N_gvm']
+    N_thr = len(meta['thr'])
+    N_atl = len(meta['atl'])
+    N_mod = len(meta['mod'])
+
+    # Normalize and split full beta vector
+    beta = normalize_beta(beta)
+    beta_atl = beta[:N_atl]
+    beta_mod = beta[N_atl:N_atl + N_mod]
+    beta_thr = beta[-N_thr:]
+
+    # Calculate linear combinations
+    gv_array = np.zeros((N_sub * N_ses, N_gvm), dtype='float')
+    for sesh in range(N_sub * N_ses):
+        gvlc = 0  # Graph-Vector Linear Combination (GVLC)
+        for atl in range(N_atl):
+            for mod in range(N_mod):
+                for thr in range(N_thr):
+                    gvlc += GVDAT[sesh][atl][mod][thr] * beta_atl[atl] * \
+                            beta_mod[mod] * beta_thr[thr]
+        gv_array[sesh] = gvlc
+    return gv_array
 
 
 if __name__ == "__main__":
@@ -253,8 +303,8 @@ if __name__ == "__main__":
     icc = True
     disc = True
 
-    #embedding_types = ['topology']
-    embedding_types = ['topology', 'OMNI', 'ASE']
+    embedding_types = ['topology']
+    #embedding_types = ['topology', 'OMNI', 'ASE']
     modalities = ['dwi', 'func']
     template = 'MNI152_T1'
     mets = ["global_efficiency", "average_clustering",
@@ -268,25 +318,24 @@ if __name__ == "__main__":
 
     sessions = ['1', '2', '3', '4', '5', '6', '7', '8', '9', '10']
 
-    subject_dict, modality_grids = make_subject_dict(modalities, base_dir,
-                                                     thr_type, mets,
-                                                     embedding_types,
-                                                     template,
-                                                     sessions)
-    sub_dict_clean = cleanNullTerms(subject_dict)
+    subject_dict_file_path = f"{base_dir}/pynets_subject_dict_topology.pkl"
 
-    subject_dict_file_path = f"{base_dir}/pynets_subject_dict.pkl"
+    if not os.path.isfile(subject_dict_file_path):
+        subject_dict, modality_grids = make_subject_dict(modalities, base_dir,
+                                                         thr_type, mets,
+                                                         embedding_types,
+                                                         template,
+                                                         sessions)
+        sub_dict_clean = cleanNullTerms(subject_dict)
 
-    with open(subject_dict_file_path, 'wb') as f:
-        dill.dump(sub_dict_clean, f)
-    f.close()
+        with open(subject_dict_file_path, 'wb') as f:
+            dill.dump(sub_dict_clean, f)
+        f.close()
+    else:
+        with open(subject_dict_file_path, 'rb') as f:
+            sub_dict_clean = dill.load(f)
+        f.close()
 
-    # with open(subject_dict_file_path, 'rb') as f:
-    #     sub_dict_clean = dill.load(f)
-    # f.close()
-
-    embedding_types = ['topology']
-    modalities = ['func']
     ix = 0
     for modality in modalities:
         hyperparams = eval(f"hyperparams_{modality}")
@@ -357,8 +406,8 @@ if __name__ == "__main__":
                             df_wide = pd.DataFrame(id_dict).T
                             df_wide = df_wide.add_prefix(f"{met}_visit_")
                             try:
-                                print('Cronbach Alpha...')
                                 c_alpha = pg.cronbach_alpha(data=df_wide)
+                                print('Cronbach Alpha...')
                             except:
                                 print('FAILED...')
                                 print(df_wide)
@@ -389,8 +438,7 @@ if __name__ == "__main__":
                     Y = np.array(id_list)
                     bad_ixs = [i[1] for i in np.argwhere(np.isnan(X_top))]
                     for m in set(bad_ixs):
-                        if (X_top.shape[0] - bad_ixs.count(m)
-                        ) / X_top.shape[0] < 0.50:
+                        if (X_top.shape[0] - bad_ixs.count(m)) / X_top.shape[0] < 0.50:
                             X_top = np.delete(X_top, m, axis=1)
                     imp = IterativeImputer(max_iter=50, random_state=42)
                     X_top = imp.fit_transform(X_top)
